+++ infra/postgres/migrations/0047_implicit_feedback.down.sql
@@ -1,0 +1,4 @@
+BEGIN;
+DROP TABLE IF EXISTS implicit_feedback;
+DELETE FROM schema_migration WHERE version='0047_implicit_feedback';
+COMMIT;