+++ packages/application/src/index.ts
@@ -22,6 +22,7 @@ export * from './goal-evaluator.js';
 export * from './goal-input-inference.js';
 export * from './workflow-controller.js';
 export * from './goal-service.js';
+export * from './implicit-feedback.js';
 export * from './goal-cancellation-service.js';
 export * from './goal-patch-service.js';
 export * from './workflow-revision.js';